# contracts/openapi/discovery_api.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Discover Egypt API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.discoveregypt.com/v1"
    }
  ],
  "paths": {
...
    "/listings": {
      "get": {
        "operationId": "listListings",
        "parameters": [
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "city", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedListingResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ListingSummary": {
        "type": "object",
        "required": ["id", "category", "title", "city", "location", "price", "rating", "reviewCount", "cancellationPolicy", "vendorName"],
        "properties": {
          "id": { "type": "string" },
          "category": { "type": "string" },
          "title": { "type": "string" },
          "city": { "type": "string" },
          "location": { "type": "string" },
          "price": { "type": "number" },
          "rating": { "type": "number" },
          "reviewCount": { "type": "integer" },
          "cancellationPolicy": { "type": "string" },
          "vendorName": { "type": "string" }
        }
      },
...
      "PaginatedListingResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ListingSummary" }
          },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "totalCount": { "type": "integer" }
        }
      },
...
    }
  }
}